Send request for actions with and without fields
